+++ src/game/tests.rs
@@ -555,6 +555,7 @@ mod vintage_cube_magda;
 mod vintage_cube_malcolm;
 mod vintage_cube_malevolent_rumble;
 mod vintage_cube_mana;
+mod vintage_cube_mana_drain;
 mod vintage_cube_mana_tithe;
 mod vintage_cube_manamorphose;
 mod vintage_cube_manifold_key;